requirements.txt
flask==2.3.3
pillow==10.0.0
opencv-python==4.8.0.74
pytesseract==0.3.10
numpy==1.24.3
requests>=2.32.2
python-dotenv==1.0.0
pdf2image==1.16.3
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0 